/Fix Pay Window(final)

+++ FootballFieldManagement/FootballFieldManagement/ViewModels/PayViewModel.cs
@@ -147,7 +147,7 @@ namespace FootballFieldManagement.ViewModels
                 var good = GoodsDAL.Instance.GetGood(billInfo.IdGoods.ToString());
                 note += good.Name + " x " + billInfo.Quantity.ToString() + "\t" + (good.UnitPrice * billInfo.Quantity).ToString() + Environment.NewLine;
             }
-            Bill bill = new Bill(int.Parse(parameter.txbIdBill.Text), CurrentAccount.IdAccount, DateTime.Now, DateTime.Now, DateTime.Now, 1, double.Parse(parameter.txbDiscount.Text), double.Parse(parameter.txbSumOfPrice.Text), int.Parse(parameter.txbIdFieldInfo.Text), note);
+            Bill bill = new Bill(int.Parse(parameter.txbIdBill.Text), CurrentAccount.IdAccount, DateTime.Now, DateTime.Now, DateTime.Now, 1, long.Parse(parameter.txbDiscount.Text), long.Parse(parameter.txbSumOfPrice.Text), int.Parse(parameter.txbIdFieldInfo.Text), note);
             if (BillDAL.Instance.UpdateOnDB(bill))
             {
                 foreach (var billInfo in billInfos)
